Refactor apollo client setup into separate file
